requirements.txt
asgiref==3.8.1
astroid==3.1.0
attrs==23.2.0
autoflake==2.3.1
autopep8==2.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.1
crispy-bootstrap4==2024.1
cryptography==42.0.7
cssbeautifier==1.15.1
defusedxml==0.8.0rc2
dill==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djlint==1.34.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.5.1
EditorConfig==0.12.4
exceptiongroup==1.2.1
Faker==25.2.0
gunicorn==22.0.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.7
inflection==0.5.1
isort==5.13.2
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pathspec==0.12.1
phonenumbers==8.13.36
pillow==10.3.0
platformdirs==4.2.1
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pylint==3.1.1
pylint-plugin-utils==0.8.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.35.1
regex==2023.12.25
requests==2.32.1
requests-oauthlib==2.0.0
rpds-py==0.18.1
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wsproto==1.2.0
